# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateTaskV2Payload": {
        "properties": {
          "animalId": {
            "minimum": 1,
            "nullable": true,
            "type": "integer"
          },
          "enclosureId": {
            "minimum": 1,
            "nullable": true,
            "type": "integer"
          },
          "repeatIntervHours": {
            "minimum": 1,
            "type": "integer"
          },
          "taskDesc": {
            "type": "string"
          },
          "taskName": {
            "type": "string"
          }
        },
        "required": [
          "repeatIntervHours",
          "taskDesc",
          "taskName"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "example": "admin access required",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Task": {
        "properties": {
          "complete": {
            "type": "boolean"
          },
          "lastCompleted": {
            "type": "string"
          },
          "repeatIntervHours": {
            "type": "integer"
          },
          "taskDesc": {
            "type": "string"
          },
          "taskId": {
            "type": "integer"
          },
          "taskName": {
            "type": "string"
          }
        },
        "required": [
          "complete",
          "lastCompleted",
          "repeatIntervHours",
          "taskDesc",
          "taskId",
          "taskName"
        ],
        "type": "object"
      },
      "TaskCompletionResponse": {
        "properties": {
          "tasksReset": {
            "type": "integer"
          }
        },
        "required": [
          "tasksReset"
        ],
        "type": "object"
      },
      "TaskWithSubject": {
        "properties": {
          "animalId": {
            "nullable": true,
            "type": "integer"
          },
          "complete": {
            "type": "boolean"
          },
          "enclosureId": {
            "nullable": true,
            "type": "integer"
          },
          "lastCompleted": {
            "type": "string"
          },
          "repeatIntervHours": {
            "type": "integer"
          },
          "taskDesc": {
            "type": "string"
          },
          "taskId": {
            "type": "integer"
          },
          "taskName": {
            "type": "string"
          }
        },
        "required": [
          "animalId",
          "complete",
          "enclosureId",
          "lastCompleted",
          "repeatIntervHours",
          "taskDesc",
          "taskId",
          "taskName"
        ],
        "type": "object"
      },
...
      "UpdateTaskV2Payload": {
        "properties": {
          "animalId": {
            "minimum": 1,
            "nullable": true,
            "type": "integer"
          },
          "complete": {
            "type": "boolean"
          },
          "enclosureId": {
            "minimum": 1,
            "nullable": true,
            "type": "integer"
          },
          "lastCompleted": {
            "type": "string"
          },
          "repeatIntervHours": {
            "minimum": 1,
            "type": "integer"
          },
          "taskDesc": {
            "type": "string"
          },
          "taskName": {
            "type": "string"
          }
        },
        "required": [
          "lastCompleted",
          "repeatIntervHours",
          "taskDesc",
          "taskName"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Raw JWT token from POST /users/login. Sent verbatim with no \"Bearer \" prefix.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "REST API for managing animals, enclosures and care tasks.\n\nOnly v2 routes are documented here. v1 remains served at /api/v1 for\nbackwards compatibility but is not part of this contract.",
    "title": "Animal Family API",
    "version": "2.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/tasks": {
      "get": {
        "description": "Pass animalId or enclosureId to return only the tasks attached to that subject. Supplying both is rejected.",
        "operationId": "listTasks",
        "parameters": [
          {
            "description": "Only tasks attached to this animal",
            "in": "query",
            "name": "animalId",
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Only tasks attached to this enclosure",
            "in": "query",
            "name": "enclosureId",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskWithSubject"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List the caller's tasks",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "description": "A task belongs to exactly one subject, so supply either animalId or enclosureId. The subject must belong to the caller.",
        "operationId": "createTask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskV2Payload"
              }
            }
          },
          "description": "Task to create",
          "required": true,
          "x-originalParamName": "task"
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/check-completion": {
      "get": {
        "description": "Intended for a scheduler. Resets tasks whose repeat interval has elapsed and sends their notifications in the background.",
        "operationId": "checkTaskCompletion",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCompletionResponse"
                }
              }
            },
            "description": "OK"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Reset repeating tasks that are due",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{id}": {
      "delete": {
        "operationId": "deleteTask",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete one of the caller's tasks",
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "getTask",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get one of the caller's tasks",
        "tags": [
          "tasks"
        ]
      },
      "put": {
        "description": "Also used to mark a task complete or incomplete, and to move it to a different subject. Leaving both subject fields null keeps the current subject.",
        "operationId": "updateTask",
        "parameters": [
          {
            "description": "Task ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskV2Payload"
              }
            }
          },
          "description": "Updated task fields",
          "required": true,
          "x-originalParamName": "task"
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update one of the caller's tasks",
        "tags": [
          "tasks"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Relative to the API host, configured per environment.",
      "url": "/api/v2"
    }
  ]
}